+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## 5.52.0 - 2020-03-09
 ### Added
 - Quest enemy data from Rayshift to nice quest phase.
+- `phasesWithEnemies` to raw and nice quest.
 
 ## 5.51.0 - 2020-03-07
 ### Added